# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  company: string;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  email: string;
  phone: string;
  company: string;
}

export interface ClientList {
  data: Client[];
  total: number;
  page: number;
  pageSize: number;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  pending: "pending",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  name: string;
  clientId: number;
  clientName: string;
  status: ProjectStatus;
  startDate: string;
  estimatedRevenue: number;
  createdAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  active: "active",
  pending: "pending",
  completed: "completed",
} as const;

export interface CreateProjectBody {
  name: string;
  clientId: number;
  status: CreateProjectBodyStatus;
  startDate: string;
  estimatedRevenue: number;
}

export interface ProjectList {
  data: Project[];
  total: number;
  page: number;
  pageSize: number;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  estimatedPrice: number;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  description: string;
  estimatedPrice: number;
}

export interface DashboardSummary {
  totalClients: number;
  activeProjects: number;
  completedProjects: number;
  estimatedRevenue: number;
}

export interface ProjectsByMonth {
  month: string;
  count: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  client_added: "client_added",
  project_created: "project_created",
  project_updated: "project_updated",
  service_added: "service_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  createdAt: string;
}

export interface RevenueByMonth {
  month: string;
  revenue: number;
}

export interface ProjectsByStatus {
  status: string;
  count: number;
}

export type ListClientsParams = {
  search?: string;
  page?: number;
  pageSize?: number;
};

export type ListProjectsParams = {
  search?: string;
  status?: ListProjectsStatus;
  page?: number;
  pageSize?: number;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  active: "active",
  pending: "pending",
  completed: "completed",
} as const;