# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MBS Dashboard API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clients
    description: Client operations
  - name: projects
    description: Project operations
  - name: services
    description: Service operations
  - name: reports
    description: Report operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientList"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, pending, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary metrics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/projects-by-month:
    get:
      operationId: getProjectsByMonth
      tags: [dashboard]
      summary: Get projects count grouped by month
      responses:
        "200":
          description: Projects by month data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectsByMonth"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /reports/revenue-by-month:
    get:
      operationId: getRevenueByMonth
      tags: [reports]
      summary: Get revenue grouped by month
      responses:
        "200":
          description: Revenue by month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueByMonth"

  /reports/projects-by-status:
    get:
      operationId: getProjectsByStatus
      tags: [reports]
      summary: Get projects grouped by status
      responses:
        "200":
          description: Projects by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectsByStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - company
        - createdAt

    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
      required:
        - name
        - email
        - phone
        - company

    ClientList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - data
        - total
        - page
        - pageSize

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        status:
          type: string
          enum: [active, pending, completed]
        startDate:
          type: string
          format: date
        estimatedRevenue:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - clientId
        - clientName
        - status
        - startDate
        - estimatedRevenue
        - createdAt

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: integer
        status:
          type: string
          enum: [active, pending, completed]
        startDate:
          type: string
          format: date
        estimatedRevenue:
          type: number
      required:
        - name
        - clientId
        - status
        - startDate
        - estimatedRevenue

    ProjectList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - data
        - total
        - page
        - pageSize

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        estimatedPrice:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - estimatedPrice
        - createdAt

    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        estimatedPrice:
          type: number
      required:
        - name
        - description
        - estimatedPrice

    DashboardSummary:
      type: object
      properties:
        totalClients:
          type: integer
        activeProjects:
          type: integer
        completedProjects:
          type: integer
        estimatedRevenue:
          type: number
      required:
        - totalClients
        - activeProjects
        - completedProjects
        - estimatedRevenue

    ProjectsByMonth:
      type: object
      properties:
        month:
          type: string
        count:
          type: integer
      required:
        - month
        - count

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [client_added, project_created, project_updated, service_added]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - createdAt

    RevenueByMonth:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
      required:
        - month
        - revenue

    ProjectsByStatus:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count