# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all clients
 */
export const listClientsQueryPageDefault = 1;
export const listClientsQueryPageSizeDefault = 10;

export const ListClientsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  page: zod.coerce.number().default(listClientsQueryPageDefault),
  pageSize: zod.coerce.number().default(listClientsQueryPageSizeDefault),
});

export const ListClientsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      email: zod.string(),
      phone: zod.string(),
      company: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  pageSize: zod.number(),
});

/**
 * @summary Create a new client
 */
export const CreateClientBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  company: zod.string(),
});

/**
 * @summary Get a client by ID
 */
export const GetClientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetClientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  company: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all projects
 */
export const listProjectsQueryPageDefault = 1;
export const listProjectsQueryPageSizeDefault = 10;

export const ListProjectsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.enum(["active", "pending", "completed"]).optional(),
  page: zod.coerce.number().default(listProjectsQueryPageDefault),
  pageSize: zod.coerce.number().default(listProjectsQueryPageSizeDefault),
});

export const ListProjectsResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      clientId: zod.number(),
      clientName: zod.string(),
      status: zod.enum(["active", "pending", "completed"]),
      startDate: zod.coerce.date(),
      estimatedRevenue: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  pageSize: zod.number(),
});

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  clientId: zod.number(),
  status: zod.enum(["active", "pending", "completed"]),
  startDate: zod.coerce.date(),
  estimatedRevenue: zod.number(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  clientId: zod.number(),
  clientName: zod.string(),
  status: zod.enum(["active", "pending", "completed"]),
  startDate: zod.coerce.date(),
  estimatedRevenue: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all services
 */
export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  estimatedPrice: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Create a new service
 */
export const CreateServiceBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  estimatedPrice: zod.number(),
});

/**
 * @summary Get dashboard summary metrics
 */
export const GetDashboardSummaryResponse = zod.object({
  totalClients: zod.number(),
  activeProjects: zod.number(),
  completedProjects: zod.number(),
  estimatedRevenue: zod.number(),
});

/**
 * @summary Get projects count grouped by month
 */
export const GetProjectsByMonthResponseItem = zod.object({
  month: zod.string(),
  count: zod.number(),
});
export const GetProjectsByMonthResponse = zod.array(
  GetProjectsByMonthResponseItem,
);

/**
 * @summary Get recent activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum([
    "client_added",
    "project_created",
    "project_updated",
    "service_added",
  ]),
  description: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);

/**
 * @summary Get revenue grouped by month
 */
export const GetRevenueByMonthResponseItem = zod.object({
  month: zod.string(),
  revenue: zod.number(),
});
export const GetRevenueByMonthResponse = zod.array(
  GetRevenueByMonthResponseItem,
);

/**
 * @summary Get projects grouped by status
 */
export const GetProjectsByStatusResponseItem = zod.object({
  status: zod.string(),
  count: zod.number(),
});
export const GetProjectsByStatusResponse = zod.array(
  GetProjectsByStatusResponseItem,
);