# lib/api-zod/src/generated/types/createClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClientBody {
  name: string;
  email: string;
  phone: string;
  company: string;
}