# lib/api-zod/src/generated/types/projectList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from "./project";

export interface ProjectList {
  data: Project[];
  total: number;
  page: number;
  pageSize: number;
}