# lib/api-zod/src/generated/types/projectsByMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectsByMonth {
  month: string;
  count: number;
}