# lib/api-zod/src/generated/types/projectsByStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectsByStatus {
  status: string;
  count: number;
}