# lib/api-zod/src/generated/types/revenueByMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBS Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevenueByMonth {
  month: string;
  revenue: number;
}